requirements.txt
requests
Flask
python-dotenv
reverse-geocoder
numpy
matplotlib
duckdb
tenacity
mapbox_vector_tile
aiohttp
